make correct

+++ controllers/user.controller.js
@@ -1,8 +1,9 @@
+import { mongo } from "mongoose";
 import User from "../models/user.model.js";
 import generateToken from "../utils/generateToken.js";
 
 // Create a new user
-const createUser = async (req, res) => {
+const createUser = async (req, res, next) => {
   try {
     const { firstName, lastName, email, password, role } = req.body;
 
@@ -40,7 +41,7 @@ const createUser = async (req, res) => {
       throw new Error("Invalid user data");
     }
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
@@ -59,21 +60,21 @@ const loginUser = async (req, res) => {
 };
 
 //logout user
-const logoutUser = async (req, res) => {
+const logoutUser = async (req, res, next) => {
   try {
     res.cookie("jwt", "", {
       httpOnly: true,
-      expires: new Date(0), // expire immediately
+      expires: new Date(0),
     });
 
     res.status(200).json({ success: true, message: "Logged out successfully" });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 //Get all users
-const getUsers = async (req, res) => {
+const getUsers = async (req, res, next) => {
   try {
     const users = await User.find();
 
@@ -83,12 +84,12 @@ const getUsers = async (req, res) => {
       data: users,
     });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 // Get a user by ID
-const getUserById = async (req, res) => {
+const getUserById = async (req, res, next) => {
   try {
     const user = await User.findById(req.params.id);
 
@@ -100,20 +101,23 @@ const getUserById = async (req, res) => {
 
     res.status(200).json({ success: true, data: user });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 // Edit a users details
-const editUser = async (req, res) => {
+const editUser = async (req, res, next) => {
   try {
     const { firstName, lastName, email, password, role } = req.body;
 
-    const updatedUser = await User.findByIdAndUpdate(
-      req.params.id,
-      { firstName, lastName, email, password, role },
-      { new: true }
-    );
+    // Corrected comparison: convert ObjectId to a string
+    if (req.user._id.toString() === req.params.id) {
+      const updatedUser = await User.findByIdAndUpdate(
+        req.params.id,
+        { firstName, lastName, email, password, role },
+        { new: true }
+      );
+    }
 
     if (!updatedUser) {
       return res
@@ -127,32 +131,41 @@ const editUser = async (req, res) => {
       data: updatedUser,
     });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 // Delete a user
-const deleteUser = async (req, res) => {
+const deleteUser = async (req, res, next) => {
+  // console.log(req.user);
+  // console.log(req.params.id);
   try {
-    const deletedUser = await User.findByIdAndDelete(req.params.id);
+    // const deletedUser = await User.findByIdAndDelete(req.params.id);
+    if (req.user._id.toString() === req.params.id) {
+      const deletedUser = await User.findByIdAndDelete(req.params.id);
 
-    if (!deletedUser) {
-      return res
-        .status(404)
-        .json({ success: false, message: "User not found" });
-    }
+      if (!deletedUser) {
+        return res
+          .status(404)
+          .json({ success: false, message: "User not found" });
+      }
 
-    res.status(200).json({
-      success: true,
-      message: "User deleted successfully",
-      data: deletedUser,
-    });
+      res.status(200).json({
+        success: true,
+        message: "User deleted successfully",
+        data: deletedUser,
+      });
+    } else
+      res.status(404).json({
+        success: false,
+        message: "You are not authorized to delete this user",
+      });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
-const getUserbyName = async (req, res) => {
+const getUserbyName = async (req, res, next) => {
   try {
     const user = await User.find({
       $or: [{ firstName: "David" }, { lastName: "Daniels" }],
@@ -165,11 +178,13 @@ const getUserbyName = async (req, res) => {
     }
 
     res.json(user);
-  } catch (error) {}
+  } catch (error) {
+    next(error);
+  }
 };
 
 //search for a user by name
-const searchUsersByName = async (req, res) => {
+const searchUsersByName = async (req, res, next) => {
   try {
     const name = req.query.name;
 
@@ -193,12 +208,12 @@ const searchUsersByName = async (req, res) => {
       data: users,
     });
   } catch (error) {
-    res.status(500).json({ success: false, message: error.message });
+    next(error);
   }
 };
 
 // Search for users by name better version
-const searchUsersByName2 = async (req, res) => {
+const searchUsersByName2 = async (req, res, next) => {
   try {
     // Get the search term from query parameters
     const { name } = req.query;
@@ -248,10 +263,7 @@ const searchUsersByName2 = async (req, res) => {
     });
   } catch (error) {
     console.error("Search user error:", error);
-    res.status(500).json({
-      success: false,
-      message: "An error occurred while searching for users",
-    });
+    next(error);
   }
 };
 
@@ -265,4 +277,5 @@ export {
   searchUsersByName,
   loginUser,
   logoutUser,
+  searchUsersByName2,
 };
